+++ data-raw/data_stations.R
@@ -1,99 +1,69 @@
 # Stations contains information about known monitoring stations
 
 stations <- list(
-    "BBLW" = list(
-        code = "BBLW",
+    "BBLW" = ace_station(
         name = "Ballard Bridge",
-        latitude = NA,
-        longitude = NA,
-        columns = c("Time",
-                    "SaltA",
-                    "SaltB",
-                    "SaltC",
-                    "TempA",
-                    "TempB",
-                    "TempC"),
-        depths = c("A" = 11,
-                   "B" = 21,
-                   "C" = 32)
+        code = "BBLW",
+        data_columns = c("Time",
+                         "SaltA",
+                         "SaltB",
+                         "SaltC",
+                         "TempA",
+                         "TempB",
+                         "TempC"),
+        depths = c("A" = 11, "B" = 21, "C" = 32)
     ),
-    "FBLW" = list(
-        code = "FBLW",
+    "FBLW" = ace_station(
         name = "Fremont Bridge",
-        latitude = NA,
-        longitude = NA,
-        columns = c("Time",
-                    "SaltA",
-                    "SaltB",
-                    "SaltC",
-                    "TempA",
-                    "TempB",
-                    "TempC"),
-        depths = c("A" = 18,
-                   "B" = 31,
-                   "C" = 40)
+        code = "FBLW",
+        data_columns = c("Time",
+                         "SaltA",
+                         "SaltB",
+                         "SaltC",
+                         "TempA",
+                         "TempB",
+                         "TempC"),
+        depths = c("A" = 18, "B" = 31, "C" = 40)
     ),
-    "UBLW" = list(
-        code = "UBLW",
+    "UBLW" = ace_station(
         name = "University Bridge",
-        latitude = NA,
-        longitude = NA,
-        columns = c("Time",
-                    "SaltA",
-                    "SaltB",
-                    "SaltC",
-                    "TempA",
-                    "TempB",
-                    "TempC"),
-        depths = c("A" = 8,
-                   "B" = 21,
-                   "C" = 35)
+        code = "UBLW",
+        data_columns = c("Time",
+                         "SaltA",
+                         "SaltB",
+                         "SaltC",
+                         "TempA",
+                         "TempB",
+                         "TempC"),
+        depths = c("A" = 8, "B" = 21, "C" = 35)
     ),
-    "GWLW" = list(
-        code = "GWLW",
+    "GWLW" = ace_station(
         name = "Gas Works Park",
-        latitude = NA,
-        longitude = NA,
-        columns = c(
-            "Time",
-            "SaltA",
-            "SaltB",
-            "SaltC",
-            "SaltD",
-            "TempA",
-            "TempB",
-            "TempC",
-            "TempD"
-        ),
-        depths = c(
-            "A" = 5,
-            "B" = 14,
-            "C" = 29,
-            "D" = 44
-        )
+        code = "GWLW",
+        data_columns = c("Time",
+                         "SaltA",
+                         "SaltB",
+                         "SaltC",
+                         "SaltD",
+                         "TempA",
+                         "TempB",
+                         "TempC",
+                         "TempD"),
+        depths = c("A" = 5, "B" = 14, "C" = 29, "D" = 44)
     ),
-    "LLLW" = list(
-        code = "LLLW",
+    "LLLW" = ace_station(
         name = "Large Lock",
-        latitude = NA,
-        longitude = NA,
-        columns = c(
-            "Time",
-            "SaltA",
-            "SaltB",
-            "SaltC",
-            "SaltD",
-            "TempA",
-            "TempB",
-            "TempC",
-            "TempD"
-        ),
-        depths = c(
-            "A" = 18,
-            "B" = 28,
-            "C" = 36,
-            "D" = 43
-        )
+        code = "LLLW",
+        data_columns = c("Time",
+                         "SaltA",
+                         "SaltB",
+                         "SaltC",
+                         "SaltD",
+                         "TempA",
+                         "TempB",
+                         "TempC",
+                         "TempD"),
+        depths = c("A" = 18, "B" = 28, "C" = 36, "D" = 43)
     )
 )